pyproject.toml
[project]
name = "taxonomise"
version = "0.1.0"
description = "Semantic taxonomy classification for document corpora"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "IGL" }
]
keywords = ["taxonomy", "classification", "nlp", "embeddings", "semantic-similarity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "sentence-transformers>=2.2",
    "spacy>=3.5",
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "joblib>=1.3",
]

[project.optional-dependencies]
keywords = [
    "rake-nltk>=1.0",
    "yake>=0.4",
    "keybert>=0.8",
    "requests>=2.28",
]
zeroshot = [
    "transformers>=4.30",
    "torch>=2.0",
    "accelerate>=0.20",
]
all = [
    "taxonomise[keywords,zeroshot]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
taxonomise = "taxonomise.cli:cli"

[project.urls]
Homepage = "https://github.com/igl/taxonomise"
Documentation = "https://github.com/igl/taxonomise#readme"
Repository = "https://github.com/igl/taxonomise"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/taxonomise"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
